requirements.txt
pdf2image
python-docx
comtypes